% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(file_name = rlang::.data)
}
\arguments{
\item{file_name}{You can insert the file_name and path to reach the tsv file or pipe it with a read_delim.}
}
\value{
A DataFrame version of file_name. Have in mind, I have only cleaning few columns:
        EQ_PRIMARY, LOCATION_NAME, DATE, and TOTAL_DEATHS.
}
\description{
You can use this function in two ways. Assigning the file_name (e.g. ../path/my_file.txt) with path or piping it
as sequence of a read_delim.
}
\examples{

\dontrun{
# Loading using a third party function and then cleaning teh DataFrame.
readr::read_delim("inst/extdata/signif.txt",
                  delim = "\\t") \%>\%
                          eq_clean_data()

# Piping a DataFrame.
any_dataframe \%>\% eq_clean_data()

# Assigning a file_name
eq_clean_data(file_name = "my_folder/signif.txt")}

}
