% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df_2 = rlang::.data)
}
\arguments{
\item{df_2}{A DataFrame with LOCATION_NAME column.}
}
\value{
Adds a new column to the DataFrame called LOCATION, following the specifications:
        Title Case and without country name.
}
\description{
This function is a intermediate step to clean the raw data. The `eq_clean_data` will use it behind
the scenes to insert the column LOCATION.
}
\examples{

\dontrun{
# Piping a DataFrame with LOCATION_NAME to be converted in LOCATION.
readr::read_delim("inst/extdata/signif.txt",
                  delim = "\\t") \%>\%
                      eq_location_clean()}

}
