% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(df_map = rlang::.data, annot_col)
}
\arguments{
\item{df_map}{A DataFrame with DATE, LONGITUDE, LATITUDE, and EQ_PRIMARY columns.}

\item{annot_col}{The information to be displayed inside of the popup.}
}
\value{
A map with earthquakes as circles.
}
\description{
This function plots a straightforward map using circles to determine the eartquake location. The radius
of the circle is according to the magnitude (in Richter Scale) of the earthquake.
}
\examples{

\dontrun{
# Assuing the my_dataset has DATE, LONGITUDE, LATITUDE, and EQ_PRIMARY columns.
my_dataset \%>\% eq_map(annot_col = "DATE")

# Full example.
readr::read_delim("inst/extdata/signif.txt",
                 delim = "\\t") \%>\%
                     eq_clean_data() \%>\%
                         dplyr::filter(COUNTRY == "MEXICO" &
                                       lubridate::year(DATE) >= 2000) \%>\%
                                           eq_map(annot_col = "DATE")}

}
