% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomTimeline
}
\value{
Creates the visuals to be plotted by the geom_timeline.
}
\description{
This functions is the Geom in charge to creates the visuals. You do not need to use the geom_timeline
to have the benefits of this function as you can see in the examples. Just uses the `ggplot2::layer`.
}
\examples{

\dontrun{
# Creating a Dataset.
df_example <- eq_clean_data(file_name = 'inst/extdata/signif.txt') \%>\%
    dplyr::filter(YEAR > 2010,
                  COUNTRY \%in\% c("CHINA",
                                 "JAPAN")) \%>\%
            dplyr::select(DATE,
                          COUNTRY,
                          EQ_PRIMARY,
                          TOTAL_DEATHS,
                          LOCATION)

# Plotting using layer.
ggplot2::ggplot() +
    ggplot2::layer(geom = GeomTimeline,
                   mapping = aes(x = df_example$DATE,
                                 y = df_example$COUNTRY,
                                 size = df_example$EQ_PRIMARY,
                                 color = df_example$TOTAL_DEATHS),
                   data = df_example,
                   stat = 'identity',
                   position = 'identity',
                   show.legend = NA,
                   inherit.aes = TRUE,
                   params = list(na.rm = FALSE))}

}
\keyword{datasets}
