% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimelineLabel.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomTimelineLabel
}
\value{
Plots the n_max most deadly earthquakes from a given dataset.
}
\description{
This function creates a line from the center of the top deadly earthquake and label it with the location.
}
\examples{

\dontrun{
# Creating a Dataset.
df_example <- eq_clean_data(file_name = 'inst/extdata/signif.txt') \%>\%
    dplyr::filter(YEAR > 2010,
                  COUNTRY \%in\% c("CHINA",
                                 "JAPAN")) \%>\%
            dplyr::select(DATE,
                          COUNTRY,
                          EQ_PRIMARY,
                          TOTAL_DEATHS,
                          LOCATION)


df_example \%>\%
   ggplot2::ggplot() +
       geom_timeline(aes(x = DATE,
                         y = COUNTRY,
                         size = EQ_PRIMARY,
                         color = TOTAL_DEATHS)) +
           ggplot2::layer(geom = GeomTimelineLabel,
                          mapping = aes(x = DATE,
                                        label = LOCATION,
                                        y = COUNTRY,
                                        mag = EQ_PRIMARY,
                                        color = TOTAL_DEATHS),
                          stat = 'identity',
                          position = 'identity',
                          show.legend = NA,
                          inherit.aes = TRUE,
                          params = list(na.rm = FALSE))}

}
\keyword{datasets}
