% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_msdr.R
\name{theme_msdr}
\alias{theme_msdr}
\title{theme_msdr}
\usage{
theme_msdr()
}
\value{
Remove the gray background, the grid, the panel border, etc. Turns the
        plot clean.
}
\description{
I used this theme to change the overall visuals of the ggplot2 to became more likely to the
examples.
}
\examples{

\dontrun{
## Full example of geom_timeline using eq_clean_data function to load the raw data and clean it.

# Loading and cleaning. Later subsetting the data.
df_earth <- eq_clean_data(file_name = 'inst/extdata/signif.txt') \%>\%
    dplyr::filter(YEAR > 2010,
                  COUNTRY \%in\% c("CHINA","JAPAN")) \%>\%
                       dplyr::select(DATE,
                       COUNTRY,
                       EQ_PRIMARY,
                       TOTAL_DEATHS,
                       LOCATION) \%>\%

    ggplot2::ggplot() +
        geom_timeline(aes(x = DATE,
                          y = COUNTRY,
                          size = EQ_PRIMARY,
                          color = TOTAL_DEATHS)) +

    theme_msdr()}

}
